Refactor the main.cc file and move certain files to a cli subdirectory.
